third_party/blink/web_tests/external/wpt/WebCryptoAPI/import_export/okp_importKey_Ed25519.https.any-expected.txt

This is a testharness.js-based test.
Found 20 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [verify])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [verify])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [verify])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [verify])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [verify, verify])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), {name: Ed25519}, true, [verify, verify])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [verify, verify])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(kty, crv, x), Ed25519, true, [verify, verify])
  assert_equals: Correct number of JWK members expected 6 but got 5
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(crv, d, x, kty), {name: Ed25519}, true, [sign])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(crv, d, x, kty), {name: Ed25519}, true, [sign])
  assert_equals: Correct number of JWK members expected 7 but got 6
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(crv, d, x, kty), Ed25519, true, [sign])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(crv, d, x, kty), Ed25519, true, [sign])
  assert_equals: Correct number of JWK members expected 7 but got 6
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(crv, d, x, kty), {name: Ed25519}, true, [sign, sign])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(crv, d, x, kty), {name: Ed25519}, true, [sign, sign])
  assert_equals: Correct number of JWK members expected 7 but got 6
[FAIL] Good parameters with JWK alg Ed25519: Ed25519 (jwk, object(crv, d, x, kty), Ed25519, true, [sign, sign])
  assert_unreached: Threw an unexpected error: DataError: The JWK "alg" member was inconsistent with that specified by the Web Crypto call Reached unreachable code
[FAIL] Good parameters with JWK alg EdDSA: Ed25519 (jwk, object(crv, d, x, kty), Ed25519, true, [sign, sign])
  assert_equals: Correct number of JWK members expected 7 but got 6
Harness: the test ran to completion.
